Added bulk upload feature for assets
