+++ app/Http/Controllers/Api/ImportController.php
@@ -176,7 +176,7 @@ class ImportController extends Controller
         $redirectTo = 'hardware.index';
         switch ($request->get('import-type')) {
             case 'asset':
-//            case 'asset_bulk':
+            case 'asset_bulk':
                 $redirectTo = 'hardware.index';
                 break;
             case 'accessory':
