+++ app/Http/Livewire/Importer.php
@@ -239,8 +239,8 @@ class Importer extends Component
             'item_name' => trans('general.item_name_var', ['item' => trans('general.asset')]),
             'asset_tag' => trans('general.asset_tag'),
             'asset_model' => trans('general.model_name'),
+            'asset_type' => trans('general.asset_type'),
             'status' => trans('general.status'),
-            'category' => trans('general.category'),
             'asset_notes' => trans('general.item_notes', ['item' => trans('admin/hardware/general.asset')]),
             'quantity' => trans('general.quantity'),
         ];
